More logical placement refactoring
