+++ packages/node/src/events/install/operation.ts
@@ -1,0 +1,23 @@
+import { Address, Node } from "@counterfactual/types";
+
+import { ProposedAppInstanceInfo } from "../../models";
+import { Store } from "../../store";
+import { getChannelFromPeerAddress } from "../../utils";
+
+export async function setAppInstanceIDForProposeInstall(
+  selfAddress: Address,
+  store: Store,
+  params: Node.InterNodeProposeInstallParams
+) {
+  const channel = await getChannelFromPeerAddress(
+    selfAddress,
+    params.peerAddress,
+    store
+  );
+  const proposedAppInstance = new ProposedAppInstanceInfo(
+    params.appInstanceId,
+    params
+  );
+
+  await store.addAppInstanceProposal(channel, proposedAppInstance);
+}
