+++ packages/node/src/methods/app-instance/get-state/controller.ts
@@ -1,0 +1,23 @@
+import { Node } from "@counterfactual/types";
+
+import { RequestHandler } from "../../../request-handler";
+import { ERRORS } from "../../errors";
+
+import { getAppInstanceState } from "./operation";
+
+/**
+ * Handles the retrieval of an AppInstance's state.
+ * @param this
+ * @param params
+ */
+export default async function getAppInstanceStateController(
+  this: RequestHandler,
+  params: Node.GetStateParams
+): Promise<Node.GetStateResult> {
+  if (!params.appInstanceId) {
+    Promise.reject(ERRORS.NO_APP_INSTANCE_ID_FOR_GET_STATE);
+  }
+  return {
+    state: await getAppInstanceState(params.appInstanceId, this.store)
+  };
+}
