+++ packages/node/src/methods/index.ts
@@ -1,21 +1,19 @@
-import {
-  getAppInstanceStateController,
-  getInstalledAppInstancesController,
-  getProposedAppInstancesController
-} from "./app-instance-operations";
-import installAppInstanceController from "./install/controller";
-import proposeInstallVirtualAppInstanceController from "./propose-install-virtual/controller";
-import { proposeInstallAppInstanceController } from "./propose-install/controller";
+import getInstalledAppInstancesController from "./app-instance/get-all/controller";
+import getAppInstanceStateController from "./app-instance/get-state/controller";
+import installAppInstanceController from "./app-instance/install/controller";
+import proposeInstallVirtualAppInstanceController from "./app-instance/propose-install-virtual/controller";
+import { proposeInstallAppInstanceController } from "./app-instance/propose-install/controller";
+import getProposedAppInstancesController from "./proposed-app-instance/get-all/controller";
 import createMultisigController from "./state-channel/create/controller";
 import getAllChannelAddressesController from "./state-channel/get-all/controller";
 
 export {
-  installAppInstanceController,
-  proposeInstallAppInstanceController,
   createMultisigController,
   getAllChannelAddressesController,
   getAppInstanceStateController,
   getInstalledAppInstancesController,
   getProposedAppInstancesController,
+  installAppInstanceController,
+  proposeInstallAppInstanceController,
   proposeInstallVirtualAppInstanceController
 };
