+++ packages/node/src/methods/proposed-app-instance/get-all/controller.ts
@@ -1,0 +1,11 @@
+import { Node } from "@counterfactual/types";
+
+import { RequestHandler } from "../../../request-handler";
+
+export default async function getProposedAppInstancesController(
+  this: RequestHandler
+): Promise<Node.GetAppInstancesResult> {
+  return {
+    appInstances: await this.store.getProposedAppInstances()
+  };
+}
